CORDA-3960: Move ThreadDumpUtils.kt to `core-utils`
